+++ src/ProtectedRoute.jsx
@@ -2,7 +2,7 @@ import React, { useEffect } from 'react';
 import PropTypes from 'prop-types';
 import useLoginStore from './stores/loginStore';
 import useModalStore from './stores/modalStore';
-import LoginModal from './components/LoginModal';
+import LoginModal from './pages/LoginPage/LoginModal';
 
 function ProtectedRoute({ children }) {
   const { loggedIn } = useLoginStore(state => ({
