+++ src/pages/DetailPage/HotDetailPage.jsx
