+++ src/pages/DetailPage/NewDetailPage.jsx
