+++ src/pages/DetailPage/SellerDetailPage.jsx
@@ -1,63 +1,65 @@
 import React, { useEffect, useState } from 'react';
-import { useLocation, useNavigate } from 'react-router-dom';
+import { useNavigate, useParams } from 'react-router-dom';
 import axios from 'axios';
 import InfiniteScroll from 'react-infinite-scroll-component';
-import ItemDetail from '../../components/ItemDetail.jsx';
+import ItemDetail from '../../components/ItemDetail';
 
-function SearchDetailPage() {
-  const location = useLocation();
+function SellerDetailPage() {
   const API_URL = import.meta.env.VITE_APP_URL;
   const navigate = useNavigate();
-  const searchParams = new URLSearchParams(location.search);
-  const keyword = searchParams.get('keyword');
+  const { auctionId } = useParams();
 
   const [searchResults, setSearchResults] = useState([]);
+  const [nickname, setNickname] = useState('');
   const [page, setPage] = useState(0);
   const [hasMore, setHasMore] = useState(true);
+  const [isLoading, setIsLoading] = useState(false);
 
   useEffect(() => {
-    setSearchResults([]);
-    setPage(0);
     const fetchSearchResults = async () => {
+      setIsLoading(true);
       try {
         const response = await axios.get(
-          `${API_URL}/api/auction/search?keyword=${keyword}&page=${page}&size=20`,
+          `${API_URL}/api/auction/${auctionId}/seller?page=${page}&size=8`,
         );
 
         if (response.data.status === 'Success') {
           setSearchResults(prevResults => [
             ...prevResults,
-            ...response.data.data.items,
+            ...response.data.data.auctionList,
           ]);
+          setNickname(response.data.data.nickname);
           setHasMore(page < response.data.data.totalPage - 1);
         }
       } catch (error) {
         console.error('Error while fetching search results:', error);
+      } finally {
+        setIsLoading(false);
       }
     };
 
     fetchSearchResults();
-  }, [keyword, page]);
+  }, [page]);
 
   const loadMoreData = () => {
     setPage(prevPage => prevPage + 1);
   };
 
-  const goToProductDetail = auctionId => {
-    navigate(`/auction/${auctionId}`);
+  const goToProductDetail = id => {
+    navigate(`/auction/${id}`);
   };
 
   return (
     <div className="whitespace-nowrap max-w-screen-lg mx-auto p-4">
-      <div className="text-3xl font-extrabold m-4 text-deepblue2">
+      <div className="text-lg mb-4 text-deepblue2">
         {' '}
-        {keyword} 와 관련된 상품은요...
+        {nickname} 님이 판매하고 있는 상품은요...
       </div>
       <InfiniteScroll
         dataLength={searchResults.length}
         next={loadMoreData}
         hasMore={hasMore}
-        loader={<h4>불러오는 중...</h4>}
+        loader={isLoading && <h4>불러오는 중...</h4>}
         endMessage={<p>마지막 페이지입니다.</p>}
       >
         <div className="grid grid-cols-4 gap-4">
@@ -74,4 +76,4 @@ function SearchDetailPage() {
   );
 }
 
-export default SearchDetailPage;
+export default SellerDetailPage;
