+++ src/pages/MainPage/ImageSlider.jsx
