+++ src/pages/MainPage/slidesData.js
@@ -1,0 +1,31 @@
+import bannerMain from '../../assets/banner-main.jpg';
+import bannerHot from '../../assets/banner-hot.jpg';
+import bannerDeadline from '../../assets/banner-deadline.jpg';
+import bannerNew from '../../assets/banner-new.jpg';
+import bannerUpload from '../../assets/banner-upload.jpg';
+
+const slidesData = [
+  {
+    id: 0,
+    url: bannerMain,
+  },
+  {
+    id: 1,
+    url: bannerHot,
+  },
+  {
+    id: 2,
+    url: bannerDeadline,
+  },
+
+  {
+    id: 3,
+    url: bannerNew,
+  },
+  {
+    id: 4,
+    url: bannerUpload,
+  },
+];
+
+export default slidesData;
