+++ src/pages/MyPage/MyShoppingPage/PurchaseRecord.jsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import axios from 'axios';
-import useLoginStore from '../stores/loginStore';
+import useLoginStore from '../../../stores/loginStore';
 import CustomDatePicker from './CustomDatePicker';
 import { PurchaseItem } from './PurchaseItem';
 import { PurchaseContainer } from './ShoppingContainer';
